:hammer: refactor file logic
