+++ api/utils/adapters.py
